--- a/src/lib.rs
+++ b/src/lib.rs
@@ -68,6 +68,7 @@ pub use crate::validator::ValidatingParserConfig;
 
 pub use crate::operators_validator::OperatorValidatorConfig;
 pub use crate::operators_validator::WasmModuleResources;
+pub use crate::operators_validator::DefaultWasmModuleResources;
 
 pub use crate::readers::CodeSectionReader;
 pub use crate::readers::CustomSectionContent;