--- a/src/lib.rs
+++ b/src/lib.rs
@@ -66,9 +66,9 @@ pub use crate::validator::ValidatingOperatorParser;
 pub use crate::validator::ValidatingParser;
 pub use crate::validator::ValidatingParserConfig;
 
+pub use crate::operators_validator::DefaultWasmModuleResources;
 pub use crate::operators_validator::OperatorValidatorConfig;
 pub use crate::operators_validator::WasmModuleResources;
-pub use crate::operators_validator::DefaultWasmModuleResources;
 
 pub use crate::readers::CodeSectionReader;
 pub use crate::readers::CustomSectionContent;