--- a/src/lib.rs
+++ b/src/lib.rs
@@ -66,13 +66,19 @@ pub use crate::validator::ValidatingOperatorParser;
 pub use crate::validator::ValidatingParser;
 pub use crate::validator::ValidatingParserConfig;
 
+pub use crate::module_resources::WasmFuncType;
+pub use crate::module_resources::WasmGlobalType;
+pub use crate::module_resources::WasmMemoryType;
+pub use crate::module_resources::WasmModuleResources;
+pub use crate::module_resources::WasmTableType;
+pub use crate::module_resources::WasmType;
+
+pub(crate) use crate::module_resources::{
+    wasm_func_type_inputs,
+    wasm_func_type_outputs,
+};
+
 pub use crate::operators_validator::OperatorValidatorConfig;
-pub use crate::operators_validator::WasmFuncType;
-pub use crate::operators_validator::WasmGlobalType;
-pub use crate::operators_validator::WasmMemoryType;
-pub use crate::operators_validator::WasmModuleResources;
-pub use crate::operators_validator::WasmTableType;
-pub use crate::operators_validator::WasmType;
 
 pub use crate::readers::CodeSectionReader;
 pub use crate::readers::CustomSectionContent;
@@ -118,6 +124,7 @@ pub use crate::readers::TypeSectionReader;
 
 mod binary_reader;
 mod limits;
+mod module_resources;
 mod operators_validator;
 mod parser;
 mod primitives;