--- a/src/lib.rs
+++ b/src/lib.rs
@@ -73,10 +73,7 @@ pub use crate::module_resources::WasmModuleResources;
 pub use crate::module_resources::WasmTableType;
 pub use crate::module_resources::WasmType;
 
-pub(crate) use crate::module_resources::{
-    wasm_func_type_inputs,
-    wasm_func_type_outputs,
-};
+pub(crate) use crate::module_resources::{wasm_func_type_inputs, wasm_func_type_outputs};
 
 pub use crate::operators_validator::OperatorValidatorConfig;