This PR extends the `WasmModuleResources` trait in order to make it possible for 3rd party crates to use their own types while reusing the validation infrastructure provided by the `wasmparser` crate.

The core of this change is reflected by the changes to the `WasmModuleResources` trait:

## Overview

These are the new validation traits introduced by this PR:

![2020-01-16-122736_1320x241_scrot](https://user-images.githubusercontent.com/8193155/72521408-a20a8800-385b-11ea-9c3f-4aa1389a71a2.png)

## Old Interface

- Uses slices
- Uses `wasmparser` types directly

```rust
pub trait WasmModuleResources {
    fn types(&self) -> &[FuncType];
    fn tables(&self) -> &[TableType];
    fn memories(&self) -> &[MemoryType];
    fn globals(&self) -> &[GlobalType];
    fn func_type_indices(&self) -> &[u32];
    fn element_count(&self) -> u32;
    fn data_count(&self) -> u32;
}
```

## New Interface

- Associated type allow customization points
- No longer depends on slices as interfacing type
- We could even go further with iterators but I wanted to implement this as a practical next step (incremental)

```rust
pub trait WasmModuleResources {
    /// The function type used for validation.
    type FuncType: WasmFuncType;
    /// The table type used for validation.
    type TableType: WasmTableType;
    /// The memory type used for validation.
    type MemoryType: WasmMemoryType;
    /// The global type used for validation.
    type GlobalType: WasmGlobalType;

    /// Returns the type at given index.
    fn type_at(&self, at: u32) -> Option<&Self::FuncType>;
    /// Returns the table at given index if any.
    fn table_at(&self, at: u32) -> Option<&Self::TableType>;
    /// Returns the linear memory at given index.
    fn memory_at(&self, at: u32) -> Option<&Self::MemoryType>;
    /// Returns the global variable at given index.
    fn global_at(&self, at: u32) -> Option<&Self::GlobalType>;
    /// Returns the function signature ID at given index.
    fn func_type_id_at(&self, at: u32) -> Option<u32>;

    /// Returns the number of elements.
    fn element_count(&self) -> u32;
    /// Returns the number of bytes in the Wasm data section.
    fn data_count(&self) -> u32;
}
```

## Benchmarks

tl;dr: No change in performance detected.

`master`:

```
     Running target/release/deps/benchmark-de6568819adee6d2
it works benchmark      time:   [1.7750 us 1.8170 us 1.8710 us]                               
                        change: [-9.1930% -2.5691% +3.5327%] (p = 0.47 > 0.05)
                        No change in performance detected.
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) high mild
  8 (8.00%) high severe

validator no fails benchmark                                                                             
                        time:   [5.5755 us 5.7014 us 5.8478 us]
                        change: [+0.2051% +3.4391% +7.0131%] (p = 0.06 > 0.05)
                        No change in performance detected.
Found 13 outliers among 100 measurements (13.00%)
  5 (5.00%) high mild
  8 (8.00%) high severe

validate benchmark      time:   [5.5023 us 5.6307 us 5.7774 us]                                
                        change: [-2.1047% +0.6309% +3.8226%] (p = 0.70 > 0.05)
                        No change in performance detected.
Found 14 outliers among 100 measurements (14.00%)
  3 (3.00%) high mild
  11 (11.00%) high severe
```

`generalize-validation`:

```
     Running target/release/deps/benchmark-de6568819adee6d2
it works benchmark      time:   [1.7702 us 1.8144 us 1.8714 us]                               
                        change: [-5.6447% -2.2495% +1.0744%] (p = 0.21 > 0.05)
                        No change in performance detected.
Found 15 outliers among 100 measurements (15.00%)
  2 (2.00%) high mild
  13 (13.00%) high severe

validator no fails benchmark                                                                             
                        time:   [5.8344 us 6.0047 us 6.2036 us]
                        change: [-0.4388% +3.6992% +8.1534%] (p = 0.08 > 0.05)
                        No change in performance detected.
Found 18 outliers among 100 measurements (18.00%)
  4 (4.00%) high mild
  14 (14.00%) high severe

validate benchmark      time:   [5.4699 us 5.5826 us 5.7223 us]                                
                        change: [-3.7465% -0.3642% +3.1391%] (p = 0.84 > 0.05)
                        No change in performance detected.
Found 14 outliers among 100 measurements (14.00%)
  5 (5.00%) high mild
  9 (9.00%) high severe
```

## Future Steps

This PR introduced the `WasmType` trait that has a compatibility method allowing user code to convert their custom Wasm types into Wasm type the `wasmparser` crate understands. This was needed because the `wasmparser` crate stores those types into some data structures during validation.

For a future PR we might be able to remove this compatibility method and make the whole validation generic over some user defined custom Wasm types. However, I am not sure if the changes required for this are actually worth the efforts which is why I did not include those changes into this PR.

The `WasmType` trait and its compatibility method `to_parser_type`:

```rust
/// Types that quality as Wasm types for validation purposes.
///
/// Must be comparable with `wasmparser` given Wasm types and
/// must be comparable to themselves.
pub trait WasmType: PartialEq<crate::Type> + PartialEq + Eq {
    /// Converts the custom type into a `wasmparser` known type.
    fn to_parser_type(&self) -> crate::Type;
}
```