@sunfishcode 2020-01-15
Interesting! One thing to keep in mind here is that an alternative to this approach would be to have third-party users just create their own `ValidatingParserResources` (wasmparser could make its version `pub` if it'd be useful) by copying data from their own data structures into it. That obviously has some overhead, but it's only handful of dynamic allocations and likely not that much total memory compared to the size of a corresponding wasm module.

A big question about the patch is, what is the performance impact of this patch? First, we have a some benchmarks in the repository that can be run with `cargo bench`, which we should make sure don't regress. The tests used there are all small wasm modules, so it'd also be good to measure the performance on some larger wasm modules as well.

Also, we need to be careful about dynamic allocations, which can cause performance effects that are sometimes harder to observe outside of larger applications. At a first pass, it appears there are a few new dynamic allocations here, in `wasm_func_type_inputs` and `wasm_func_type_outputs`. How difficult would it be to avoid those?