@Robbepop 2020-01-15
Benchmark Results

`cargo bench` on `master` branch:

```
     Running target/release/deps/benchmark-8378c2fd07a0c47c
Benchmarking it works benchmark: Collecting 100 samples in estimated 7.5852 s (10k iterati                                                                                          it works benchmark      time:   [1.7471 us 1.7801 us 1.8467 us]
                        change: [-4.8360% -1.4652% +1.8791%] (p = 0.42 > 0.05)
                        No change in performance detected.
Found 15 outliers among 100 measurements (15.00%)
  3 (3.00%) low mild
  1 (1.00%) high mild
  11 (11.00%) high severe

Benchmarking validator no fails benchmark: Collecting 100 samples in estimated 10.702 s (5                                                                                          validator no fails benchmark                        
                        time:   [5.4536 us 5.5048 us 5.5700 us]
                        change: [-9.7212% -6.5739% -2.7280%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 16 outliers among 100 measurements (16.00%)
  3 (3.00%) high mild
  13 (13.00%) high severe

Benchmarking validate benchmark: Collecting 100 samples in estimated 10.120 s (5050 iterat                                                                                          validate benchmark      time:   [5.3918 us 5.5025 us 5.6512 us]
                        change: [-5.6111% -2.6055% +0.7406%] (p = 0.12 > 0.05)
                        No change in performance detected.
Found 13 outliers among 100 measurements (13.00%)
  5 (5.00%) high mild
  8 (8.00%) high severe
```

`cargo bench` on `generalize-validation` branch:

```
     Running target/release/deps/benchmark-de6568819adee6d2
it works benchmark      time:   [1.7655 us 1.8133 us 1.8751 us]                               
                        change: [-6.7002% -2.4573% +2.0317%] (p = 0.29 > 0.05)
                        No change in performance detected.
Found 17 outliers among 100 measurements (17.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  13 (13.00%) high severe

validator no fails benchmark                                                                             
                        time:   [5.9293 us 6.0483 us 6.1927 us]
                        change: [-3.5594% -0.7874% +2.1262%] (p = 0.59 > 0.05)
                        No change in performance detected.
Found 16 outliers among 100 measurements (16.00%)
  2 (2.00%) high mild
  14 (14.00%) high severe

validate benchmark      time:   [5.5433 us 5.6119 us 5.6970 us]                                
                        change: [-6.8912% -3.2146% +0.0741%] (p = 0.08 > 0.05)
                        No change in performance detected.
Found 14 outliers among 100 measurements (14.00%)
  4 (4.00%) high mild
  10 (10.00%) high severe
```

So it looks like the performance regression exists but is minimal.
Also note that this benchmarks the dynamically allocating `wasm_func_type_inputs` and `wasm_func_type_outputs` so maybe we can get better results once that has been fixed.