@Robbepop 2020-01-15
I just removed the dynamic allocations of the `wasm_func_type_{inputs,outputs}` functions and can see our expected performance improvements:

```
     Running target/release/deps/benchmark-de6568819adee6d2
it works benchmark      time:   [1.7781 us 1.7956 us 1.8187 us]                               
                        change: [-8.4591% -3.6018% +1.3921%] (p = 0.16 > 0.05)
                        No change in performance detected.
Found 17 outliers among 100 measurements (17.00%)
  2 (2.00%) high mild
  15 (15.00%) high severe

validator no fails benchmark                                                                             
                        time:   [5.9056 us 6.1645 us 6.4688 us]
                        change: [+0.9319% +4.7262% +8.9154%] (p = 0.02 < 0.05)
                        Change within noise threshold.
Found 17 outliers among 100 measurements (17.00%)
  5 (5.00%) high mild
  12 (12.00%) high severe

validate benchmark      time:   [5.4440 us 5.5382 us 5.6632 us]                                
                        change: [-14.655% -8.1688% -3.0888%] (p = 0.01 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  9 (9.00%) high mild
  6 (6.00%) high severe
```